finalised work

+++ app/Http/Livewire/Trip/RemoveWorker.php
@@ -3,13 +3,15 @@
 namespace App\Http\Livewire\Trip;
 
 use App\Models\Trip;
+use App\Models\User;
 use App\Models\UserTrip;
 use Livewire\Component;
 
-class Delete extends Component
+class RemoveWorker extends Component
 {
     public $showModal = false;
     public Trip $trip;
+    public User $user;
 
     protected $listeners = [
         'load'=>'load'
@@ -17,25 +19,26 @@ class Delete extends Component
 
     /**
      * @param Trip $trip
+     * @param User $user
      */
-    public function load(Trip $trip){
+    public function load(Trip $trip, User $user){
         $this->trip = $trip;
+        $this->user = $user;
         $this->showModal = true;
     }
 
     public function delete(){
         foreach(UserTrip::all() as $triping){
-            if($triping->trip_id === $this->trip->id){
+            if($triping->trip_id === $this->trip->id && $triping->user_id === $this->user->id){
                 $triping->delete();
             }
         }
-        $this->trip->delete();
         $this->emit("success", "trip deleted successfully!");
-        $this->emit("tripDeleted");
+        $this->emit("userAdded");
         $this->showModal = false;
     }
     public function render()
     {
-        return view('livewire.trip.delete');
+        return view('livewire.trip.remove-worker');
     }
 }
